duplicated code arreglado + consecuentemente se arreglan long methods

+++ src/main/java/dds/monedero/model/Deposito.java
@@ -1,10 +1,11 @@
 package dds.monedero.model;
 
+import java.math.BigDecimal;
 import java.time.LocalDate;
 
 public class Deposito extends Movimiento{
 
-  public Deposito(LocalDate fecha, double monto) {
+  public Deposito(LocalDate fecha, BigDecimal monto) {
     super(fecha, monto);
   }
 }
