+++ src/main/java/dds/monedero/model/Extraccion.java
@@ -1,11 +1,13 @@
 package dds.monedero.model;
 
+import java.math.BigDecimal;
 import java.time.LocalDate;
 
 public class Extraccion extends Movimiento {
 
-
-  public Extraccion(LocalDate fecha, double monto) {
+  public Extraccion(LocalDate fecha, BigDecimal monto) {
     super(fecha, monto);
   }
+
+
 }
